refactor(stores): projectConfigs.ts를 패키지로 승격 + 타입 추출
1,553줄 → types.ts 352줄(타입 20종) + index.ts 1,233줄. 아직 800줄 한도
초과이나 이 태스크의 예상 범위다 — 액션 72개 승격은 Task 8에서 처리한다.

인터페이스 ProjectConfigsState에 export 추가(패키지 내부 전용이지만
Task 8에서 도메인 모듈이 SetFn/GetFn 대신 직접 참조). 상대 경로 한 단계
깊어짐(../services, ../config → ../../). export type 재노출은 로컬
바인딩을 만들지 않으므로, index.ts 본문이 애노테이션으로 쓰는 타입
(ProjectInfo·SkillConfig·AgentConfig·ProjectConfigSummary·DBProject·
TabType·GlobalConfigSummary·ConfigChangeEvent)은 별도 import type으로
바인딩했다.

소비자 실측 11개 타입 + 훅 useProjectConfigsStore 재노출로 26개 소비자
파일의 import 경로 불변. 표면 스냅샷(projectConfigs.surface.test.ts) +
기존 테스트 61건(projectConfigs.test.ts) 통과, 전체 스위트 202파일
4,358건 통과, tsc/ESLint 0, build 성공.
